Return integer validators as original value
While doing spot checks of json diffs of the original template and converted
through cfn2py I noticed all of the integer values were appearing as integer
values while the original was always strings. This change returns the
original value and reduces the json diff to only be boolean changes.