Tries to convert JSON structures into troposphere functions/objects
The general case is when the function has the same name as the key.
If not, 3 other quirk cases are handled (depending on the type of
value in function_quirks dict (respectively string, list, dict):
- The function has a different name.
- The key expects an array of function calls.
- The function requires a mapping of input values for its arguments.

I added the values for the available examples (except for the
OpenStack ones and the one depending on awacs). Many additional
cases should handled by adding some values in some tables, though
probably not all cases.

It needs an additional patch to add imports for the additional
function calls.