Allow an optional resource_prefix in Template
I find myself doing "MyStack" + <resource_name> really often in my
templates. It seemed like there would be a better way, and so I thought
I'd just try something.  I'm not 100% sold on this - it's a little
magical, but I thought I'd put it out there since it was such a simple
change and it could clean up a lot of boilerplate code.

The place where I can see this being problematic is when you later are
Ref() or GetAtt()'ing a resource that has had a prefix added to it.

Anyway, just an experiment. Be curious what people think about it.