Fix #369: Add back support for Python 2.6.9
The use of a dictionary comprehension in #354 caused support for Python 2.6.9
to be dropped. This change converts it into a simple iteration. Note: while it
would be ideal to add travis support to prevent further regressions the amount
of change in the tests is prohibitive at this time. Python 2.6.9 support is
"best effort" for now.