Validation on Day-of-month / Day-of-week fields
AWS's documentation says that Day-of-month and Day-of-week fields can not both be a value or *.

Here we apply this check in the validator and use a the message from the documentation.

This is desirable as this is a common mistake made by those using this feature for the first time.