Update validators.py
Current code seems to assume positive_integer(x) converts to int, while it doesn't.
This problem generates 'Integer must be one of following...' error when the value is in the list because it tries to find a string in an array of ints. 

Other methods such as integer_range do cast correctly, integer_list_item should work the same.

Ran into this issue when creating a LogGroup object with RetentionInDays set to u'7' which should be valid. Passing a number also doesn't work because the expected value should be a string.