Documentation update
Updated docs to include most of the files from the main dir into the
docs, such as code of conduct, license, release, etc.  Cleaned up the
TOC.  Found sphinx-apidoc and applied it to examples, troposphere and
tests.  Also, used the Makefile index example to add sphinx-apidoc to
the build process so files in docs/apis are created automatically.

Also added dynamic versioning to the setup.py file and some additional
classifiers.