CHANGELOG.rst Formatting Fixes
Changes:

- Silenced a Sphinx error by closing an unclosed emphasis markup.
- Fixed 3.0.0 release notes formatting.

The 3.0.0 releases notes were mostly rendering as just a paragraph, or
an an incomplete list with unexpected indentation. These changes fix the
formatting so that the 3.0.0 release notes Breaking Changes, and Changes
lists are now subsections of the 3.0.0 section. The second bullet point
in the Breaking Changes section was also throwing unexpected indentation
errors. This has been resolved now, and the section is rendering as a
literal box as expected.